+++ src/index.js
@@ -11,6 +11,8 @@ const OPTIONS = {
   loadSidebar: null,
   loadNavbar: null,
   router: false,
+  homepage: 'README.md',
+  basePath: '',
   auto2top: false
 }
 const script = document.currentScript || [].slice.call(document.getElementsByTagName('script')).pop()
@@ -33,8 +35,9 @@ let cacheRoute = null
 let cacheXhr = null
 
 const mainRender = function (cb) {
-  const route = getRoute()
+  const route = OPTIONS.basePath + getRoute()
   if (cacheRoute === route) return cb()
+
   let wait
   let basePath = cacheRoute = route
 
@@ -44,13 +47,18 @@ const mainRender = function (cb) {
     basePath = basePath.match(/(\S*\/)[^\/]+$/)[1]
   }
 
+  let page
+  if (!route) {
+    page = OPTIONS.homepage || 'README.md'
+  } else if (/\/$/.test(route)) {
+    page = `${route}README.md`
+  } else {
+    page = `${route}.md`
+  }
+
   cacheXhr && cacheXhr.abort && cacheXhr.abort()
   // Render markdown file
-  cacheXhr = load(
-    (!route || /\/$/.test(route)) ? `${route}README.md` : `${route}.md`,
-    'GET',
-    render.renderLoading)
-
+  cacheXhr = load(page, 'GET', render.renderLoading)
   cacheXhr.then(result => {
     render.renderArticle(result)
     if (OPTIONS.loadSidebar) {